KrishiBot/requirements.txt
streamlit
python-dotenv
openai
requests
gTTS
streamlit-audiorecorder